Split Auth Emulator tests into multiple files.
